+++ src/drivers/device/i2c_linux.cpp
@@ -51,8 +51,7 @@ I2C::I2C(const char *name,
 	 const char *devname,
 	 int bus,
 	 uint16_t address,
-	 uint32_t frequency,
-	 int irq) :
+	 uint32_t frequency) :
 	// base class
 	CDev(name, devname),
 	// public
